requirements.txt
absl-py==1.4.0
brotlipy==0.7.0
cachetools==5.3.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fonttools==4.40.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
imageio==2.31.1
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
Markdown==3.4.3
matplotlib==3.7.1
mpmath==1.2.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.4.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchaudio==2.0.2
torchfile==0.1.0
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
tzdata==2023.3
Werkzeug==2.3.6
